[fix] base conversion

+++ src/error.c
@@ -7,7 +7,7 @@
 
 #include "palindrome.h"
 
-long int is_str_digit(char *str)
+int is_str_digit(char *str)
 {
     for (long int i = 0; str[i] != '\0'; i++) {
         if (!isdigit(str[i]))
@@ -16,7 +16,7 @@ long int is_str_digit(char *str)
     return 1;
 }
 
-long int number_error(char *str)
+int number_error(char *str)
 {
     if (atoi(str) <= 0) {
         dprintf(2, "invalid argument\n");
@@ -25,7 +25,7 @@ long int number_error(char *str)
     return 0;
 }
 
-long int check_arg(long int ac, char **av)
+long int check_arg(int ac, char **av)
 {
     for (long int i = 3; av[i]; i++) {
         if ((strcmp(av[i], "-imin") == 0 || strcmp(av[i], "-imax") == 0 ||
